pyproject.toml
[project]
name = "mcp-complex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "python-dotenv>=1.1.1",
]
